REFACTOR: extract test stack creation into factory class
slightly reduces mental load when reading tests and potentially allows for more tests to be written against the stacks in the future
