+++ source/test/aws-instance-scheduler-remote-stack.test.ts
@@ -13,51 +13,13 @@
  *  permissions and limitations under the License.                            *
  *****************************************************************************/
 
-import { SynthUtils } from '@aws-cdk/assert';
-import * as cdk from 'aws-cdk-lib';
-import instanceSchedulerStack = require('../lib/aws-instance-scheduler-remote-stack');
-import '@aws-cdk/assert/jest';
+ import { SynthUtils } from '@aws-cdk/assert';
+ import '@aws-cdk/assert/jest';
+ import {createSpokeStack} from './instance-scheduler-stack-factory';
 
-
-function getTestStack(): cdk.Stack {
-  const envEU = { account: '111111111111', region: 'eu-west-1' };
-  const app = new cdk.App();
-  const stack = new instanceSchedulerStack.AwsInstanceSchedulerRemoteStack(app, 'stack', {
-    env: envEU,
-    description: "", 
-    solutionId: "SO0030",
-    solutionTradeMarkName: 'aws-instance-scheduler',
-    solutionProvider: 'AWS Solution Development',
-    solutionBucket: "solutions",
-    solutionName: 'aws-instance-scheduler',
-    solutionVersion: 'v1.4.0'
-  })
-
-  const instanceSchedulerRemoteStackMetadata =
-  {
-    "AWS::CloudFormation::Interface": {
-      "ParameterGroups": [{
-        "Label": {
-          "default": "Account"
-        },
-        "Parameters": [
-          "InstanceSchedulerAccount"
-        ]
-      }],
-      "ParameterLabels": {
-        "InstanceSchedulerAccount": {
-          "default": "Primary account"
-        }
-      }
-    }
-  }
-  stack.templateOptions.metadata = instanceSchedulerRemoteStackMetadata
-  stack.templateOptions.templateFormatVersion = "2010-09-09"
-  return stack;
-}
-/*
- * SnapShot Testing for the AwsInstanceSchedulerStack.
- */
-test('AwsInstanceSchedulerStack snapshot test', () => {
-  expect(SynthUtils.toCloudFormation(getTestStack())).toMatchSnapshot();
-});
+ /*
+  * SnapShot Testing for the AwsInstanceSchedulerStack.
+  */
+ test('AwsInstanceSchedulerStack snapshot test', () => {
+   expect(SynthUtils.toCloudFormation(createSpokeStack())).toMatchSnapshot();
+ });
